+++ src/Exception/ExceptionInterface.php
@@ -1,0 +1,9 @@
+<?php
+namespace Phly\Http\Exception;
+
+/**
+ * Marker interface for package-specific exceptions.
+ */
+interface ExceptionInterface
+{
+}