+++ src/IncomingRequestFactory.php
@@ -2,13 +2,13 @@
 namespace Phly\Http;
 
 use Psr\Http\Message\IncomingRequestInterface;
-use Psr\Http\Message\RequestInterface;
+use Psr\Http\Message\MessageInterface;
 use stdClass;
 
 /**
  * Class for marshaling a request object from the current PHP environment.
  *
- * Logic largely refactors ZF2's Zend\Http\PhpEnvironment\Request class.
+ * Logic largely refactored from the ZF2 Zend\Http\PhpEnvironment\Request class.
  *
  * @copyright Copyright (c) 2005-2014 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
@@ -39,20 +39,28 @@ abstract class IncomingRequestFactory
         array $cookies = null,
         array $files = null
     ) {
-        $server  = $server  ?: $_SERVER;
+        $server  = self::normalizeServer($server ?: $_SERVER);
+        $headers = self::marshalHeaders($server);
+        $url     = (string) self::marshalUriFromServer($server, $headers);
+        $method  = self::get('REQUEST_METHOD', $server, 'GET');
         $query   = $query   ?: $_GET;
         $body    = $body    ?: $_POST;
         $cookies = $cookies ?: $_COOKIE;
         $files   = $files   ?: $_FILES;
+
         $request = new IncomingRequest(
+            $url,
+            $method,
+            $headers,
             'php://input',
+            $server,
             $cookies,
-            [],
             $query,
             $body,
-            $files
+            $files,
+            []
         );
-        return self::fromServer($server, $request);
+        return $request;
     }
 
     /**
@@ -60,26 +68,22 @@ abstract class IncomingRequestFactory
      *
      * @param array $server
      * @param IncomingRequestInterface $request
-     * @return IncomingRequestInterface The $request provided, only populated with values
+     * @return void
+     * @deprecated as of 0.7.0. Use fromGlobals().
+     * @throws Exception\DeprecatedMethodException on all requests.
      */
     public static function fromServer(array $server, IncomingRequestInterface $request = null)
     {
-        $server = self::normalizeServer($server);
-
-        if (! $request) {
-            $protocol = self::get('SERVER_PROTOCOL', $server, '1.1');
-            $request  = new IncomingRequest();
-            $request->setProtocolVersion($protocol);
-        }
-
-        $request->setMethod(self::get('REQUEST_METHOD', $server, 'GET'));
-        self::setHeaders($request, self::marshalHeaders($server));
-        $request->setUrl(self::marshalUri($server, $request));
-        return $request;
+        throw new Exception\DeprecatedMethodException(sprintf(
+            '%s is deprecated as of phly/http 0.7.0dev; always use fromGlobals()',
+            __METHOD__
+        ));
     }
 
     /**
      * Access a value in an array, returning a default value if not found
+     *
+     * Will also do a case-insensitive search if a case sensitive search fails.
      *
      * @param string $key
      * @param array $values
@@ -91,6 +95,33 @@ abstract class IncomingRequestFactory
         if (array_key_exists($key, $values)) {
             return $values[$key];
         }
+
+        return $default;
+    }
+
+    /**
+     * Search for a header value.
+     *
+     * Does a case-insensitive search for a matching header.
+     *
+     * If found, it is returned as a string, using comma concatenation.
+     *
+     * If not, the $default is returned.
+     * 
+     * @param string $header 
+     * @param array $headers 
+     * @param mixed $default 
+     * @return string
+     */
+    public static function getHeader($header, array $headers, $default = null)
+    {
+        $header  = strtolower($header);
+        $headers = array_change_key_case($headers, CASE_LOWER);
+        if (array_key_exists($header, $headers)) {
+            $value = is_array($headers[$header]) ? implode(', ', $headers[$header]) : $headers[$header];
+            return $value;
+        }
+
         return $default;
     }
 
@@ -142,6 +173,7 @@ abstract class IncomingRequestFactory
             if ($value && strpos($key, 'HTTP_') === 0) {
                 $name = strtr(substr($key, 5), '_', ' ');
                 $name = strtr(ucwords(strtolower($name)), ' ', '-');
+                $name = strtolower($name);
 
                 $headers[$name] = $value;
                 continue;
@@ -150,6 +182,7 @@ abstract class IncomingRequestFactory
             if ($value && strpos($key, 'CONTENT_') === 0) {
                 $name = substr($key, 8); // Content-
                 $name = 'Content-' . (($name == 'MD5') ? $name : ucfirst(strtolower($name)));
+                $name = strtolower($name);
                 $headers[$name] = $value;
                 continue;
             }
@@ -159,41 +192,39 @@ abstract class IncomingRequestFactory
     }
 
     /**
-     * Set the headers for a request.
+     * Marshal the URI from the $_SERVER array and headers
      *
-     * Injects the given request instance with the headers provided.
-     * 
-     * @param RequestInterface $request 
-     * @param array $headers 
+     * @param array $server
+     * @param MessageInterface $request
+     * @return Uri
+     * @deprecated as of 0.7.0; use marshalUriFromServer() instead.
      */
-    public static function setHeaders(RequestInterface $request, array $headers)
+    public static function marshalUri(array $server, MessageInterface $request)
     {
-        foreach ($headers as $header => $values) {
-            $request->setHeader($header, $values);
-        }
+        return self::marshalUriFromServer($server, $request->getHeaders());
     }
 
     /**
      * Marshal the URI from the $_SERVER array and headers
      *
      * @param array $server
-     * @param IncomingRequestInterface $request
+     * @param array $headers
      * @return Uri
      */
-    public static function marshalUri(array $server, IncomingRequestInterface $request)
+    public static function marshalUriFromServer(array $server, array $headers)
     {
         // URI scheme
         $scheme = 'http';
-        $https = self::get('HTTPS', $server);
+        $https  = self::get('HTTPS', $server);
         if (($https && 'off' !== $https)
-            || $request->getHeader('x-forwarded-proto') == 'https'
+            || self::getHeader('x-forwarded-proto', $headers, false) === 'https'
         ) {
             $scheme = 'https';
         }
 
         // Set the host
         $accumulator = (object) ['host' => '', 'port' => null];
-        self::marshalHostAndPort($accumulator, $server, $request);
+        self::marshalHostAndPortFromHeaders($accumulator, $server, $headers);
         $host = $accumulator->host;
         $port = $accumulator->port;
 
@@ -220,13 +251,26 @@ abstract class IncomingRequestFactory
      * Marshal the host and port from HTTP headers and/or the PHP environment
      *
      * @param array $server
-     * @param IncomingRequestInterface $request
+     * @param MessageInterface $request
+     * @return array Array with two members, host and port, at indices 0 and 1, respectively
+     * @deprecated as of 0.7.0; use marshalHostAndPortFromHeaders() instead.
+     */
+    public static function marshalHostAndPort(stdClass $accumulator, array $server, MessageInterface $request)
+    {
+        return self::marshalHostAndPortFromHeaders($accumulator, $server, $request->getHeaders());
+    }
+
+    /**
+     * Marshal the host and port from HTTP headers and/or the PHP environment
+     *
+     * @param array $server
+     * @param array $headers
      * @return array Array with two members, host and port, at indices 0 and 1, respectively
      */
-    public static function marshalHostAndPort(stdClass $accumulator, array $server, IncomingRequestInterface $request)
+    public static function marshalHostAndPortFromHeaders(stdClass $accumulator, array $server, array $headers)
     {
-        if ($request->hasHeader('host')) {
-            return self::marshalHostAndPortFromHeader($accumulator, $request);
+        if (self::getHeader('host', $headers, false)) {
+            return self::marshalHostAndPortFromHeader($accumulator, self::getHeader('host', $headers));
         }
 
         if (! isset($server['SERVER_NAME'])) {
@@ -314,12 +358,16 @@ abstract class IncomingRequestFactory
      * Marshal the host and port from the request header
      *
      * @param stdClass $accumulator
-     * @param IncomingRequestInterface $request
+     * @param string|array $host
      * @return void
      */
-    private static function marshalHostAndPortFromHeader(stdClass $accumulator, IncomingRequestInterface $request)
+    private static function marshalHostAndPortFromHeader(stdClass $accumulator, $host)
     {
-        $accumulator->host = $request->getHeader('host');
+        if (is_array($host)) {
+            $host = implode(', ', $host);
+        }
+
+        $accumulator->host = $host;
         $accumulator->port = null;
 
         // works for regname, IPv4 & IPv6